board1

+++ banking/src/banking/domain/AccountDAO.java
@@ -206,20 +206,25 @@ public class AccountDAO {
 		while(true) {
 			System.out.print("계좌 번호: ");
 			String ano = scanner.nextLine();
-		
 			if(findAccount(ano) != null) { //찾는 계좌가 있으면
-				try {
-					conn = JDBCUtil.getConnection();
-					String sql = "DELETE FROM account WHERE ano = ?";
-					pstmt = conn.prepareStatement(sql);
-					pstmt.setString(1, ano);
-					//sql 실행
-					pstmt.executeUpdate();
-					System.out.println("결과: 정상 처리되었습니다.");
-				} catch (SQLException e) {
-					e.printStackTrace();
-				} finally {
-					JDBCUtil.close(conn, pstmt);
+				System.out.println("1.OK | 2. Cancel");
+				String menuNo = scanner.nextLine();
+				if(menuNo.equals("1")) {
+					try {
+						conn = JDBCUtil.getConnection();
+						String sql = "DELETE FROM account WHERE ano = ?";
+						pstmt = conn.prepareStatement(sql);
+						pstmt.setString(1, ano);
+						//sql 실행
+						pstmt.executeUpdate();
+						System.out.println("결과: 정상 처리되었습니다.");
+					} catch (SQLException e) {
+						e.printStackTrace();
+					} finally {
+						JDBCUtil.close(conn, pstmt);
+					}
+				}else {
+					System.out.println("삭제를 취소합니다.");
 				}
 				break;
 			}else {
@@ -240,6 +245,34 @@ public class AccountDAO {
 			String ano = scanner.nextLine();
 		
 			if(findAccount(ano) != null) { //찾는 계좌가 있으면
+				
+				try {
+					conn = JDBCUtil.getConnection(); //db 연결
+					String sql = "SELECT * FROM account WHERE ano = ?";
+					pstmt = conn.prepareStatement(sql);
+					pstmt.setString(1, ano);
+					//검색 처리
+					rs = pstmt.executeQuery();
+					while(rs.next()) { //다음 데이터가 있는 동안 반복
+						//account 객체
+						ano = rs.getString("ano"); //db에서 가져온 ano 칼럼
+						String owner = rs.getString("owner"); //db에서 가져온 ano 칼럼
+						int balance = rs.getInt("balance"); //db에서 가져온 ano 칼럼
+						
+						System.out.println("1개의 계좌를 찾았습니다.");
+						System.out.print("계좌번호: " + ano + "\t");
+						System.out.print("계좌주: " + owner + "\t");
+						System.out.println("잔고: " + balance);
+						
+						
+					}
+				} catch (SQLException e) {
+					e.printStackTrace();
+				} finally {
+					JDBCUtil.close(conn, pstmt, rs);
+				}
+				
+				
 				//db 처리
 				break;
 			}else {
